Archivos .py movidos a carpeta gestionador_tareas
